Add reshape binding.
This adds code to handle ATTR_INT_LIST, which is yet untested. Actually
I started down the wrong path initially when implementing this op, and
did all that work, only to realize it was unneessary. Rather than
leaving it in a branch and forgetting about it, I think it's better to
just add now at this early stage.