Save for backward
In order to remove the unnecessary matmulT0 and matmulT1, it's necessary
to pass the transposeA and tranposeB arguments thru to the backwards
pass.

Previously we were assuming that all arguments to ops were tensors. But
this isn't the case, there are many flags and other attributes that are
used in ops, which are not necessarally tensors... Like transposeA/B.
Or for reduction ops, the keep_dims.

Furthermore, we were always keeping references to all of the arguments
and outputs of every op during the forward pass, for potential use in
backwards passes. Different ops need different arguments remembered, so
this change introduces the saveForBackward() call. This should be more
memory efficent and simplify the current matmul situation.

PyTorch does something similiar
https://github.com/pytorch/pytorch/blob/5c809de4b4663c688ebc2cd389c2c866aa22f6e5/torch/autograd/function.py#L11