Remove TODOs on addGrad and mulDivGrad.
I'm not convinced that it's a good idea to do the backwards pass for all inputs
in a single function. It might be that backwards pass for certain inputs is not
needed, which will result in unnecessary computation. For example, in matmul,
the gradients for each input require a matmul.  If one of the gradients wasn't
needed, that would be an extra matmul. It's true that currently we always
compute all of the backwards passes - but this can be optimized - if we
continue having a different BW function for each input.