Better output routing for notebook.
In the notebook if cells contain async functions, output may be
routed to the last cell instead of the one where you would expect
the async function to run.

This change does a slightly better job than before at routing this
output. (Although not perfect.) When an output function is called
(console.log, plot, or imshow), the current stack trace is examined and
looks for the string "__cell42__.js" where 42 is a cell id.  It then
looks the found id to lookup the output html element.  In cases where
the stack trace doesn't contain a reference to a cell, it outputs to the
last run cell.