Rename util.ts to utils.ts
This is actually to work around a bug in webpack. DL also has a file
called util.ts at deps/deeplearnjs/src/util.ts which is imported
in environment.ts like this:

  import * as util from './util';

It appears that this confuses webpack with our util.ts.

Rather than spending time figuring out and debugging this issue,
we just rename for now to avoid the conflict. If this issue pops up
again, I will invest some time into a real fix.