Modify notebook output caching behavior.
Previously nb output was cached in a server-side render, and
used in the browser without re-executing the cell. Now output
is cached server side, but re-executed client side (only if the user has
JS enabled, obviously).

This is to remove the hack where innerHTML of output cells was
pulled during browser-side initial page render. This optimization
is complicating development.

This also re-enables the notebook-cell-updating feature.