Add website_main.ts to give better entry point.
The onload event is the main entry point into the website and is only
set when IS_WEB is true. However we want to add some browser-only tests
which will fail if that onload event is called. This is because
document.location.pathname will resolve to static/test.html which
results in a 404 message.